Core/Networking/Packets: Audited every single Packet type and corrected some errors (details are below). All Packets now use the newly-implemented convenience functions for readability purposes. ・Renamed `writeString16()` to `writeString16WithoutStringLengthShort()` to more clearly reflect that the string length `short` must be sent separately. ・Corrected the list of which packets are server-bound and/or client-bound in PacketReader. ・If HandshakeResponsePacket is instantiated without an argument (this should never happen), it now defaults to the safe (as in, functional) default of "-" to disable authentication ・EntityTeleportPacket (0x22) was completely missing `entityId` ・SetPlayerPositionPacket (0x0D Server) used the incorrect type for `yaw` and `pitch` (should be `float`, but were incorrectly set to `int`) ・PlayerPositionAndLookPacket (0x0D Client) used the incorrect type for `yaw` and `pitch` (should be `float`, but were incorrectly set to `int`) ・PlayerLookPacket (0x0C) used the incorrect type for `yaw` and `pitch` (should be `float`, but were incorrectly set to `int`) ・AnimationPacket (0x12) now has a writePacket() response. ・CloseWindowPacket (0x65) now has a writePacket() response. ・CollectItemPacket (0x16) now has a writePacket() response. It was previously erroneously assigned readPacket() functionality, which has since been removed. ・UseBedPacket (0x11) now has a writePacket() response. It was previously erroneously assigned readPacket() functionality, which has since been removed. ・UpdateSignPacket (0x82) now has readPacket() functionality.

+++ src/PHPCraft/Core/Networking/Packets/PlayerLookPacket.php
@@ -12,8 +12,8 @@ class PlayerLookPacket {
 	public $onGround;
 
 	public function readPacket(StreamWrapper $StreamWrapper) {
-		$this->yaw = $StreamWrapper->readInt();
-		$this->pitch = $StreamWrapper->readInt();
+		$this->yaw = $StreamWrapper->readFloat();
+		$this->pitch = $StreamWrapper->readFloat();
 		$this->onGround = $StreamWrapper->readBool();
 	}
 }
